Update the EstimatorConfig
